+++ test/test_system/test_sbipa.py
@@ -22,15 +22,16 @@
 
 from __future__ import annotations
 
-from pathlib import Path
 from test.test_system.base import ModeTest
+
+from cssfinder.examples import Example
 
 
 class Test_SBiPa(ModeTest):  # noqa: N801  # Underscore used for readability.
     """Test behavior of SBiPa mode of Gilbert algorithm."""
 
     EXPECTED_MINIMAL_NUMBER_OF_CORRECTIONS: int = 10
-    PROJECT_PATH = Path.cwd() / "examples" / "proj"
+    PROJECT_PATH = Example.proj.get_path()
     TEST_TASK_NAME: str = "test_sbipa_proj"
 
     OUT_STATE_ROW_COUNT: int = 9